+++ CORA.Rcheck/tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(CORA)
-
-test_check("CORA")